+++ src/main/java/by/epam/lab/issuetracker/interfaces/IBuildDAO.java
@@ -9,7 +9,7 @@ public interface IBuildDAO{
 
 	public List<Build> getAll() throws DAOException;	
 	public List<Build> getAll(long projectId) throws DAOException;
-	public Build get(int id) throws DAOException;	
+	public Build get(long id) throws DAOException;	
 	public void update(Build build) throws DAOException;	
 	public Build add(Build build) throws DAOException;
 }
